+++ src/gemm.h
@@ -1,5 +1,6 @@
 #ifndef GEMM_H
 #define GEMM_H
+#include<stdint.h>
 
 void gemm_bin(int M, int N, int K, float ALPHA, 
         char  *A, int lda, 
@@ -12,10 +13,10 @@ void gemm(int TA, int TB, int M, int N, int K, float ALPHA,
                     float BETA,
                     float *C, int ldc);
 void gemm_q(int TA, int TB, int M, int N, int K, float ALPHA, 
-                float *A, int lda, 
-                float *B, int ldb,
+                int8_t *A, int lda, 
+                int8_t *B, int ldb,
                 float BETA,
-                int *C, int ldc);
+                int32_t *C, int ldc);
 
 void gemm_cpu(int TA, int TB, int M, int N, int K, float ALPHA, 
         float *A, int lda, 
@@ -23,10 +24,10 @@ void gemm_cpu(int TA, int TB, int M, int N, int K, float ALPHA,
         float BETA,
         float *C, int ldc);
 void gemm_cpu_q(int TA, int TB, int M, int N, int K, float ALPHA, 
-        float *A, int lda, 
-        float *B, int ldb,
+        int8_t *A, int lda, 
+        int8_t *B, int ldb,
         float BETA,
-        int *C, int ldc);
+        int32_t *C, int ldc);
 
 #ifdef GPU
 void gemm_gpu(int TA, int TB, int M, int N, int K, float ALPHA,
