+++ src/im2col.h
@@ -1,9 +1,10 @@
 #ifndef IM2COL_H
 #define IM2COL_H
+#include<stdint.h>
 
-void im2col_cpu(float* data_im,
+void im2col_cpu(int8_t* data_im,
         int channels, int height, int width,
-        int ksize, int stride, int pad, float* data_col);
+        int ksize, int stride, int pad, int8_t* data_col);
 
 #ifdef GPU
